First sprint
It still does not work, but the most of the code is there.

The output binary must be analyzed, to see how to tweak compilation
options and programming style in order to obtain a shellcode.

Also, a script to obtain a shellcode from the binary must be developed,
and also a loader of the shellcode (as a testbed).