Fixed libc loading and fnv1a32 macro expansion
The macro expansion at compile-time has been replaced with the actual
constants, since it was too compiler-dependant. Anyway, this can always be
improved, by defining which compiler should be used. The hard-coded
constants are a temporary solution.

About the libc loading, the biggest problem was about the structure used
by fstat syscall. I couldn't find the correct one, but I found out that
it changes across versions and architectures, so at last I decided to
have a bigger size for the mmap, big enough to handle the libc library
(with a hard-coded size of 16 MB). I also improved the error check on
the mmap output.