Fixed scraping of exported symbols
It searches in dynsym section. However, strlen function was removed
because it's not in that section. In the case of the kernel, for
example, the sections have different names, so the approach will be
different.

There came in another problem: the calling convention of the binary
doesn't match the convention of libc, so it must be forced. The default
calling convention in Linux is sysv, whereas libc uses cdecl. Also, our
syscall stub was developed with sysv in mind, so it can't be cdecl
globally.