Working shellcode and loading
First, it was figured out that libc can't be called after just being
mmaped: it needs more complex steps of initialization. Anyway, it's not
important, because the shellcode is injected in a process where libc is
already initialized, it just needs the base address of it.

Second, compilation options (CFLAGS) were tweaked to generate a better
result.

Third, it was implemented a custom linker, and it was used with linking
options (LDFLAGS): it makes sure that the _start function is the first
one in the .text area.

Fourth, in the $(TARGET) objective in Makefile it was added the dumping
of the compiler version to file (it's important for the result) and the
dumping of the .text area to the file that will contain the shellcode.
So, it was implemented an "inspect" objective to analyze both the
executable and the shellcode and see if everything is fine.

Fifth, it was implemented a loader for the shellcode (test/loader.c) and
its related objective in the Makefile. The loader accepts as input the
path to the shellcode, allocates a RWX buffer to read it in, reads libc
base address from /proc/self/maps, patches the shellcode by searching a
"cookie" in it (a pattern), and finally jumps to the buffer.

Last, it was figured out that, at runtime, an ELF file gets its section
headers stripped. So, the scraping to find addresses of exported
functions can't be done on libc at runtime: the shellcode was modified
such to load the libc file in order to find function offsets in it, and
then apply those offsets to the address of the real libc, the one
properly loaded.